requirements.txt
Flask
streamlit
selenium
beautifulsoup4
pandas
plotly
requests
google-generativeai